requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.0